move region resolution to be a sibling of `region_inference`
Temporary make various fields public.
